requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pydantic
python-dotenv
aiofiles
torch
torchvision
torchaudio
opencv-python-headless
facenet-pytorch
numpy
pillow
gTTS
google-cloud-speech
google-cloud-texttospeech
